Backed Fix #4
Fixed the "Is Banned:" field of the player check report AGAIN. Was
reporting erroneous data if the argument passed was not lower-case (the
format in which data is stored in runtime by both Backend Managers.)

Also, a check of the Player object in Report.java has been implemented
to prevent future possible erroneous checks. Now, if the Player object
is null when the report code attempts to use it, the ban check method
will not execute, and instead the report will alert the user to the
error, preventing any possible miss-checks.

Plugin updated to ver. 1.2.3