Prep Work for 1.3.x
Added rudimentary command logging, a logging interface, loggers for
internal (IP-Check related) and external (other) commands, an
instantiatable date stamp class, and a new configuration option for
setting the date-stamp format for logs.

Other:
* Fixed a possible resource bug in which a for-loop was allowed to run
longer than required.
* Rewrote the checkVersion() method in Configuration.java to make adding
changes to the method easier to perform. This method still needs to be
rewritten again to prevent possible duplicate entries with-in the
configuration file and to make updating the file easier and less
painful.
* Added DEBUG boolean within Configuration to help with ease of
debugging during the update process.