Adding Features (faux 1.2.4 update)
A user requested a feature be added to the current release of the plugin
(1.2.3). This resulted in the release of "1.2.4", which was really the
1.3.x source code with it's most prominent new features disabled. The
features which were disabled have been reenabled for this commit,
though.

Faux 1.2.4 Changes:
Added permission node "ipcheck.showip". This permission is now required
by the command sender in order for IP-Check to show the IP-address of a
player check in report.

1.3.x Changes:
Internal version now reports as: "IP-Check v1.3.0_DEV-RELEASE_01."

Added options to the configuration for linking to MySQL database.
However, currently the code can only fetch these settings and store them
in memory. It currently does not use them. In other words, the fields do
nothing at the moment. Functionality will come once Baylor gets an SQL
server set up.