Rewrote Configuration Updater
Made configuration updater self-sufficient. With this, we no longer have
to tell the method to check for missing features. We don't have to tell
it anything, in fact. This method is update-proof for all future
updates, as it can detect the feature changes on it's own.

This method also performs the splitting of the configuration and
exemption information into two separate files, which will make updating
the configuration easier.