Modified Ban Command Syntax
The ban command now supports typing out a custom message. The syntax is
as follows: "/c ban <playername or ip> [message here]"

Also, the plugin now displays a "Banned" message to all players on the
server, telling them who was banned, and who banned them, in a sort of
play-off of Essentials. With this change, the ban command no longer
tells the sender what players were banned, as this new change takes over
that task.