A Plethora of Stuff...
This commit fixes a severe bug in 1.2.5 where the plugin would fail to
generate the "exempt.lst" file on first run. (This bug only effected
first time users, but regardless.)

New Additions:
Internal Error Logging - When an internal error occurs (Mainly confined
to try-catch-finally atm), a new system "ErrorLogger" will be
instantiated to write a log of the error to file. This will help
substantially in debugging future issues with the core plugin.

Custom Messages - Now, on plugin start, the plugin will choose a random
message from a list to display. Also, certain messages are hard-coded to
appear on certain days. This system also applies to the ErrorLogging
system, in that a random message is displayed at the top of every log
file (similar to that of Minecraft's own logs.)

A new Help Command - Displays help for all of the registered commands in
IP-Check.

Multiple Interface updates - Two new methods have been added to the
Backends Interface; String getRegistrar(); and int getMemorySize();
These methods are used mainly in by the ErrorLogging system. The
IpcCommand interface has received 4 new methods as well: string
getHelp(); String getSyntax(); Permission[] getPermissions(); and String
getName(); All of these are used by the new Help Command.

Number of commands is now 13.